backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.9.2
groq==0.11.0
google-generativeai==0.8.3
chromadb==0.5.5
sentence-transformers==3.0.1
torch==2.5.1
httpx==0.27.2
requests==2.32.3
pandas==2.2.3
pyyaml==6.0.2
pytest==8.3.3
